Slightly refactored code so that I didn't have to call functions from another data model in the DS model.
