+++ api/ds/dsModel.js
@@ -1,6 +1,5 @@
 const db = require('../../data/db-config');
-const { dsApi } = require('../../lib');
-const Mod = require('../mod/modModel');
+const { dsApi, dbOps, formatCohortSubmissions } = require('../../lib');
 
 /**
  * Attempts to update the complexity score of a submission.
@@ -12,41 +11,21 @@ const setComplexity = (ID, Complexity) => {
   return db('Submissions').where({ ID }).update({ Complexity });
 };
 
-const addSquad = (conn, CohortID) => {
-  return conn('Squads').insert({ CohortID }).returning('ID');
-};
-
-const addTeams = (
-  conn,
-  SquadID,
-  team1Name = 'Team 1',
-  team2Name = 'Team 2'
-) => {
-  return conn('Teams')
-    .insert([
-      { SquadID, Name: team1Name },
-      { SquadID, Name: team2Name },
-    ])
-    .returning('ID');
-};
-
-const addSingleMember = (conn, SubmissionID, TeamID) => {
-  return conn('Members').insert({ SubmissionID, TeamID }).returning('ID');
-};
-
-const addMembers = (conn, team1ID, team2ID, subIDs) => {
-  return Promise.all(
-    subIDs.map((id, i) => addSingleMember(conn, id, i < 2 ? team1ID : team2ID))
-  );
-};
-
+/**
+ * This function runs a series of transactional requests to:
+ *   1. pull and format submission data from our database
+ *   2. send the formatted data to the DSAPI for clustering
+ *   3. store the generated "squads" in our databse and put the members into teams
+ * @returns {Object} returns an object containing the new clusters (not needed)
+ */
 const clusterGeneration = () => {
   return db.transaction(async (trx) => {
     try {
       const res = {};
       const cohorts = await trx('Cohorts');
       for (let { ID } of cohorts) {
-        res[ID] = await Mod.getSubmissionsByCohort(ID);
+        const unformatted = await dbOps.getAllSubmissionsByCohort(trx, ID);
+        res[ID] = formatCohortSubmissions(unformatted);
       }
       const clusters = await dsApi.getClusters(res);
 
@@ -64,6 +43,65 @@ const clusterGeneration = () => {
       trx.rollback();
     }
   });
+};
+
+// Helper functions for storing clusters in the database
+
+/**
+ * Adds a new cluster (squad) to the database for foreign key references.
+ * @param {Object} conn a knex client instance
+ * @param {number} CohortID the integer representation of the cohort
+ * @returns {Promise} returns a promise resolving to an array with the newly generated squad ID inside
+ */
+const addSquad = (conn, CohortID) => {
+  return conn('Squads').insert({ CohortID }).returning('ID');
+};
+
+/**
+ * Adds two new teams to the database for the same squad
+ * @param {Object} conn a knex client instance
+ * @param {number} SquadID the integer representation of the squad
+ * @param {string} team1Name name of the given team (defaults to "Team #")
+ * @param {string} team2Name name of the given team (defaults to "Team #")
+ * @returns {Promise} returns a promise resolving to an array with the two new team IDs inside
+ */
+const addTeams = (
+  conn,
+  SquadID,
+  team1Name = 'Team 1',
+  team2Name = 'Team 2'
+) => {
+  return conn('Teams')
+    .insert([
+      { SquadID, Name: team1Name },
+      { SquadID, Name: team2Name },
+    ])
+    .returning('ID');
+};
+
+/**
+ * Adds a group of team members to the database
+ * @param {Object} conn a knex client instance
+ * @param {number} team1ID the integer representation of the first team
+ * @param {number} team2ID the integer representation of the second team
+ * @param {Array} subIDs an array of 4 IDs, each pointing to the submissions for the given team
+ * @returns {Promise} returns a promise resolving to an array containing all of the new members' IDs
+ */
+const addMembers = (conn, team1ID, team2ID, subIDs) => {
+  return Promise.all(
+    subIDs.map((id, i) => addSingleMember(conn, id, i < 2 ? team1ID : team2ID))
+  );
+};
+
+/**
+ * Adds a new team member to the database, should only be called by addMembers()
+ * @param {Object} conn a knex client instance
+ * @param {number} SubmissionID the integer representation of the submission
+ * @param {number} TeamID the integer representation of the team
+ * @returns {Promise} returns a promise resolving to an array with the new member's ID in it
+ */
+const addSingleMember = (conn, SubmissionID, TeamID) => {
+  return conn('Members').insert({ SubmissionID, TeamID }).returning('ID');
 };
 
 module.exports = {
