+++ lib/dsRequests.js
@@ -31,13 +31,14 @@ const submitDrawingToDS = (image) => {
   return Promise.resolve(image);
 };
 
+/**
+ * This function sends submissions to the DSAPI for clustering
+ * @param {Object} submissions a specially-formatted object that contains submissions grouped by cohort
+ * @returns {Promise} a promise that resolves to the newly clustered children
+ */
 const getClusters = (submissions) => {
-  const cohorts = Object.keys(submissions);
-  const res = {};
-  cohorts.forEach((c) => {
-    res[c] = [Object.keys(submissions[c])];
-  });
-  return Promise.resolve(res);
+  // return client.post('/clusters', submissions);
+  return Promise.resolve(submissions);
 };
 
 module.exports = {
