+++ lib/formatHelpers.js
@@ -1,4 +1,4 @@
-const formatSubmissions = (submissions) => {
+const formatCohortSubmissions = (submissions) => {
   const res = {};
   submissions.forEach((page) => {
     if (!res[page.ID]) {
@@ -53,6 +53,6 @@ const formatProfiles = (profiles) => {
 };
 
 module.exports = {
-  formatSubmissions,
+  formatCohortSubmissions,
   formatProfiles,
 };
