+++ lib/index.js
@@ -1,7 +1,8 @@
 const ops = require('./crudOps');
 const dsApi = require('./dsRequests');
-const errorHandler = require('./errorHandler');
+const dbOps = require('./dbOps');
 const oktaClient = require('./oktaClient');
+const errorHandler = require('./errorHandler');
 const uploadHelpers = require('./uploadHelpers');
 const validationCheckers = require('./validationCheckers');
 const formatters = require('./formatHelpers');
@@ -9,8 +10,9 @@ const formatters = require('./formatHelpers');
 module.exports = {
   ops,
   dsApi,
-  errorHandler,
+  dbOps,
   oktaClient,
+  errorHandler,
   uploadHelpers,
   validationCheckers,
   ...formatters,
